requirements.txt
kiutils>=1.4.0
pyyaml>=6.0
mcp>=1.0
